metodo para excluir refatorado para o padrao resst

+++ src/main/java/br/com/lucas/apirest/controllers/UsuarioController.java
@@ -7,6 +7,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -40,12 +41,17 @@ public class UsuarioController {
 		return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 	
-	@DeleteMapping(value = "excluir")
-	public ResponseEntity<String> excluir(@RequestParam(name = "idUsuario") Long idUsuario) {
+	@DeleteMapping(value = "/{idUsuario}")
+	public ResponseEntity<String> excluir(@PathVariable Long idUsuario) {
 		
-		this.usuarioRepository.deleteById(idUsuario);
+		if (idUsuario != null) {
 		
-		return new ResponseEntity<String>("Usuário Excluído com Sucesso", HttpStatus.OK);
+			this.usuarioRepository.deleteById(idUsuario);
+			
+			return new ResponseEntity<String>("Usuário Excluído com Sucesso", HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 	}
